5.1.3 Перенесем utils.js в папку /utils, переименуем в common.js и вынесем функции для рендеринга из `common.js` в `utils/render.js`
